+++ app/commands/draw.js
@@ -1,7 +1,9 @@
 import config from '../../config/index.js';
 import { COMMAND_DRAW } from '../../constants/command.js';
+import { TEXT_OK } from '../../constants/mock.js';
+import { writeHistory } from '../histories.js';
 import { PARTICIPANT_AI, PARTICIPANT_HUMAN } from '../../services/openai.js';
-import generateImage from '../../utils/generate-image.js';
+import { generateImage } from '../../utils/index.js';
 import Context from '../context.js';
 import { getPrompt, setPrompt } from '../prompts.js';
 
@@ -25,8 +27,9 @@ const execDrawCommand = async (context) => {
     .write(`\n${PARTICIPANT_AI}: `);
   try {
     const { url } = await generateImage({ prompt: context.argument, size });
-    prompt.write('OK!');
+    prompt.write(TEXT_OK);
     setPrompt(context.userId, prompt);
+    writeHistory(context.contextId, config.SETTING_AI_NAME, TEXT_OK);
     context.pushImage(url);
   } catch (err) {
     context.pushError(err);