+++ app/commands/summarize.js
@@ -1,29 +1,33 @@
-import config from '../../config/index.js';
-import { COMMAND_CONTINUE } from '../../constants/command.js';
+import { COMMAND_CONTINUE, COMMAND_SUMMARIZE } from '../../constants/command.js';
+import { t } from '../../languages/index.js';
+import { PARTICIPANT_AI, PARTICIPANT_HUMAN } from '../../services/openai.js';
 import { generateCompletion } from '../../utils/index.js';
 import MessageAction from '../actions/message.js';
 import Context from '../context.js';
-import { writeHistory } from '../histories.js';
+import { getFormattedHistory } from '../histories.js';
 import { getPrompt, setPrompt } from '../prompts.js';
 
 /**
  * @param {Context} context
  * @returns {boolean}
  */
-const isContinueCommand = (context) => context.isCommand(COMMAND_CONTINUE);
+const isSummarizeCommand = (context) => context.isCommand(COMMAND_SUMMARIZE);
 
 /**
  * @param {Context} context
  * @returns {Promise<Context>}
  */
-const execContinueCommand = async (context) => {
+const execSummarizeCommand = async (context) => {
+  const content = getFormattedHistory(context.contextId);
   const prompt = getPrompt(context.userId);
+  prompt
+    .write(`\n${PARTICIPANT_HUMAN}: `)
+    .write(`${t('__COMPLETION_SUMMARIZE_REQUEST')}\n${t('__COMPLETION_QUOTATION_MARK_OPENING')}\n${content}\n${t('__COMPLETION_QUOTATION_MARK_CLOSING')}`)
+    .write(`\n${PARTICIPANT_AI}: `);
   try {
     const { text, isFinishReasonStop } = await generateCompletion({ prompt: prompt.toString() });
-    if (!text) return context;
     prompt.write(text);
     setPrompt(context.userId, prompt);
-    writeHistory(context.contextId, config.SETTING_AI_NAME, text);
     const actions = isFinishReasonStop ? [] : [new MessageAction(COMMAND_CONTINUE)];
     context.pushText(text, actions);
   } catch (err) {
@@ -33,8 +37,6 @@ const execContinueCommand = async (context) => {
 };
 
 export {
-  isContinueCommand,
-  execContinueCommand,
+  isSummarizeCommand,
+  execSummarizeCommand,
 };
-
-export default null;