+++ tests/call.test.js
@@ -1,20 +1,18 @@
 import {
   afterEach, beforeEach, expect, test,
 } from '@jest/globals';
-import {
-  getPrompt, handleEvents, removePrompt, settings,
-} from '../app/index.js';
+import { getPrompt, handleEvents, removePrompt } from '../app/index.js';
 import config from '../config/index.js';
 import { SETTING_AI_ACTIVATED } from '../constants/setting.js';
 import storage from '../storage/index.js';
-import { createEvents, TIMEOUT, USER_ID } from './utils.js';
+import { createEvents, TIMEOUT, USER_ID_01 } from './utils.js';
 
 beforeEach(() => {
   storage.setItem(SETTING_AI_ACTIVATED, false);
 });
 
 afterEach(() => {
-  removePrompt(USER_ID);
+  removePrompt(USER_ID_01);
 });
 
 test('COMMAND_CALL', async () => {
@@ -27,7 +25,7 @@ test('COMMAND_CALL', async () => {
   } catch (err) {
     console.error(err);
   }
-  expect(getPrompt(USER_ID).lines.length).toEqual(3 * 2);
+  expect(getPrompt(USER_ID_01).lines.length).toEqual(3 * 2);
   const replies = results.map(({ messages }) => messages.map(({ text }) => text));
   expect(replies).toEqual(
     [