+++ tests/draw.test.js
@@ -3,14 +3,14 @@ import {
 } from '@jest/globals';
 import { handleEvents, getPrompt, removePrompt } from '../app/index.js';
 import { COMMAND_DRAW } from '../constants/command.js';
-import { createEvents, TIMEOUT, USER_ID } from './utils.js';
+import { createEvents, TIMEOUT, USER_ID_01 } from './utils.js';
 
 beforeEach(() => {
   //
 });
 
 afterEach(() => {
-  removePrompt(USER_ID);
+  removePrompt(USER_ID_01);
 });
 
 test('COMMAND_DRAW', async () => {
@@ -23,7 +23,7 @@ test('COMMAND_DRAW', async () => {
   } catch (err) {
     console.error(err);
   }
-  expect(getPrompt(USER_ID).lines.length).toEqual(3 * 2);
+  expect(getPrompt(USER_ID_01).lines.length).toEqual(3 * 2);
   const replies = results.map(({ messages }) => messages.map(({ originalContentUrl }) => originalContentUrl));
   expect(replies).toEqual(
     [