+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.12.0 (2022-12-30)
+
+### New Features
+
+- Implement `summarize` command
+
 ## 1.11.3 (2022-12-29)
 
 ### Bug Fixes
@@ -22,7 +28,7 @@
 
 ### New Features
 
-- Add `call` command
+- Implement `call` command
 - Add `SETTING_AI_NAME` environment variable
 - Add `SETTING_AI_ACTIVATED` environment variable
 
@@ -126,7 +132,7 @@
 
 ### Bug Fixes
 
-- Add `ja` to initial languages
+- Add `ja` initial language
 - Add `ai` alias for `chat` command
 
 ## 1.4.4 (2022-12-18)